+++ src/main/java/cn/edu/swjtu/demo/Service/UserService.java
@@ -29,7 +29,7 @@ public interface UserService {
 	// 记录用户搜索日志
 	void searchLog(String cookieid, Integer selectedBrand, Integer selectedSeries, Integer selectedClassify,
 			String selectedPaifang, Integer selectedColor, Integer selectedLichengLeft, Integer selectedLichengRight,
-			Double selectedPriceLeft, Double selectedPriceRight);
+			Integer selectedPriceLeft, Integer selectedPriceRight);
 
 	// 展示全部帖子
 	List<CarInfoWithBLOBs> showPosts();
@@ -49,7 +49,7 @@ public interface UserService {
 	// 搜索帖子
 	List<CarInfoWithBLOBs> getPostByCondition(Integer selectedBrand, Integer selectedSeries, Integer selectedClassify,
 			String selectedPaifang, Integer selectedColor, Integer selectedLichengLeft, Integer selectedLichengRight,
-			Double selectedPriceLeft, Double selectedPriceRight);
+			Integer selectedPriceLeft, Integer selectedPriceRight);
 
 	// 获取推荐
 	List<CarInfoWithBLOBs> getRecommend();
